Merry style

+++ src/main/java/com/reminiscence/reminiscence/journal/EntryController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.view.RedirectView;
 
 import java.security.Principal;
 import java.util.List;
@@ -35,7 +36,7 @@ public class EntryController {
     }
 
     @PostMapping("/entry")
-    public String createEntry(
+    public RedirectView createEntry(
             @RequestParam String body,
             Principal p
     ) {
@@ -43,19 +44,25 @@ public class EntryController {
         Entry entry = new Entry(body);
         entry.setUser(user);
         entryRepo.save(entry);
-        return "home";
+        return new RedirectView("/home");
     }
 
     @GetMapping("/entry/{id}")
     public String viewSingleEntry(
             @PathVariable long id,
-            Model model
+            Model model,
+            Principal p
     ) {
         Optional<Entry> foundEntry = entryRepo.findById(id);
         if (foundEntry.isPresent()) {
+            UserAccount user = accountRepo.findByUsername(p.getName());
             Entry entry = foundEntry.get();
-            model.addAttribute("entry", entry);
-            return "singleEntry";
+            if (user.getId() == entry.getUser().getId()) {
+                model.addAttribute("entry", entry);
+                return "singleEntry";
+            } else {
+                throw new UnauthorizedAccountException();
+            }
         }
         throw new EntryNotFoundException();
     }
@@ -63,13 +70,19 @@ public class EntryController {
     @GetMapping("/entry/{id}/update")
     public String getEntry(
             @PathVariable long id,
-            Model model
+            Model model,
+            Principal p
     ) {
         Optional<Entry> foundEntry = entryRepo.findById(id);
         if (foundEntry.isPresent()) {
+            UserAccount user = accountRepo.findByUsername(p.getName());
             Entry entry = foundEntry.get();
-            model.addAttribute("entry", entry);
-            return "updateEntry";
+            if (user.getId() == entry.getUser().getId()) {
+                model.addAttribute("entry", entry);
+                return "updateEntry";
+            } else {
+                throw new UnauthorizedAccountException();
+            }
         }
         throw new EntryNotFoundException();
     }
@@ -78,26 +91,43 @@ public class EntryController {
     public String updateEntry(
             @PathVariable long id,
             @RequestParam String body,
-            Model model
+            Model model,
+            Principal p
     ) {
         Optional<Entry> foundEntry = entryRepo.findById(id);
         if (foundEntry.isPresent()) {
+            UserAccount user = accountRepo.findByUsername(p.getName());
             Entry entry = foundEntry.get();
-            entry.setBody(body);
-            entry.setEdited(true);
-            entryRepo.save(entry);
-        } else {
-            throw new EntryNotFoundException();
+            if (user.getId() == entry.getUser().getId()) {
+                entry.setBody(body);
+                entry.setEdited(true);
+                entryRepo.save(entry);
+                return "home";
+            } else {
+                throw new UnauthorizedAccountException();
+            }
         }
-        return "home";
+        throw new EntryNotFoundException();
     }
 
     @DeleteMapping("/entry/{id}/update")
     public String deleteEntry(
-            @RequestParam long id
+            @RequestParam long id,
+            Principal p
     ) {
-        entryRepo.deleteById(id);
-        return "home";
+        Optional<Entry> foundEntry = entryRepo.findById(id);
+        if (foundEntry.isPresent()) {
+            UserAccount user = accountRepo.findByUsername(p.getName());
+            Entry entry = foundEntry.get();
+            if (user.getId() == entry.getUser().getId()) {
+                entryRepo.deleteById(id);
+                return "home";
+            } else {
+                throw new UnauthorizedAccountException();
+            }
+        } else {
+            throw new EntryNotFoundException();
+        }
     }
 
 }
